backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytz>=2023.3
python-dateutil>=2.8.0
